+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+# v0.7.1
 - add: new context menu item: create reference addons based on selected addons
 - add: new tool: File Name Fixer (this tool might help you solve the problem of missing addon files)
 - add: web proxy credential settings (issue #19)